+++ src/main/java/core/net/DownloadDialog.java
@@ -4,10 +4,12 @@ package core.net;
 import core.datatype.CBItem;
 import core.gui.HOMainFrame;
 import core.gui.RefreshManager;
+import core.gui.comp.CheckBoxTree.CheckBoxTree;
 import core.gui.comp.panel.ImagePanel;
 import core.gui.theme.ho.HOTheme;
 import core.model.HOModel;
 import core.model.HOVerwaltung;
+import core.model.UserParameter;
 import core.model.player.Player;
 import core.net.login.ProxyDialog;
 import core.util.HOLogger;
@@ -36,6 +38,8 @@ import javax.swing.JSpinner;
 import javax.swing.KeyStroke;
 import javax.swing.ListSelectionModel;
 import javax.swing.SpinnerDateModel;
+import javax.swing.border.Border;
+import javax.swing.tree.DefaultTreeModel;
 
 
 /**
@@ -53,16 +57,17 @@ public class DownloadDialog extends JDialog implements ActionListener {
 	final private JButton m_jbDownload = new JButton(hov.getLanguageString("ls.button.download"));
 	private JButton m_jbProxy = new JButton(hov.getLanguageString("ConfigureProxy"));
 	private JCheckBox m_jchOldFixtures = new JCheckBox(hov.getLanguageString("download.oldseriesdata"), false);
-	private JCheckBox m_jchOwnFixtures = new JCheckBox(hov.getLanguageString("download.currentmatches"),
-			core.model.UserParameter.instance().currentMatchlist);
-	private JCheckBox m_jchHRF = new JCheckBox(hov.getLanguageString("download.teamdata"),
-			core.model.UserParameter.instance().xmlDownload);
+	private DownloadFilter filterRoot = new DownloadFilter();
+	private CheckBoxTree downloadFilter = new CheckBoxTree();
+
+	private JCheckBox m_jchHRF = new JCheckBox(hov.getLanguageString("download.teamdata"),	core.model.UserParameter.instance().xmlDownload);
 	private JCheckBox m_jchMatchArchive = new JCheckBox(hov.getLanguageString("download.oldmatches"), false);
-	private JCheckBox m_jchFixtures = new JCheckBox(hov.getLanguageString("download.seriesdata"), core.model.UserParameter
-			.instance().fixtures);
+	private JCheckBox m_jchFixtures = new JCheckBox(hov.getLanguageString("download.seriesdata"), core.model.UserParameter.instance().fixtures);
 	private JList m_jlOldSeasons = new JList();
 	private SpinnerDateModel m_clSpinnerModel = new SpinnerDateModel();
 	private JSpinner m_jsSpinner = new JSpinner(m_clSpinnerModel);
+	private JCheckBox m_jchShowSaveDialog = new JCheckBox(hov.getLanguageString("Show_SaveHRF_Dialog"), core.model.UserParameter.instance().showHRFSaveDialog);
+
 
 	// ~ Constructors
 	// -------------------------------------------------------------------------------
@@ -123,17 +128,33 @@ public class DownloadDialog extends JDialog implements ActionListener {
 		final JPanel normalDownloadPanel = new ImagePanel(new GridLayout(3, 1, 4, 4));
 		normalDownloadPanel.setBorder(BorderFactory.createTitledBorder(hov.getLanguageString("ls.button.download")));
 
-		m_jchHRF.setToolTipText(hov.getLanguageString("download.teamdata.tt"));
-		m_jchOwnFixtures.setToolTipText(hov.getLanguageString("download.currentmatches.tt"));
-		m_jchFixtures.setToolTipText(hov.getLanguageString("download.seriesdata.tt"));
-		m_jchHRF.setOpaque(false);
-		m_jchOwnFixtures.setOpaque(false);
-		m_jchFixtures.setOpaque(false);
-		normalDownloadPanel.add(m_jchHRF);
-		normalDownloadPanel.add(m_jchOwnFixtures);
-		normalDownloadPanel.add(m_jchFixtures);
+		// Download Filter
 
-		normalDownloadPanel.setSize(200, 200);
+		final DefaultTreeModel newModel = new DefaultTreeModel(filterRoot);
+		downloadFilter.setModel(newModel);
+		newModel.reload();
+
+		// Current Matches
+		// - Official Matches
+		//    currentMatchlist selects now the node OfficialMatches.
+		//    It is the first subitem of current Matches in the Filter tree
+		downloadFilter.checkNode(filterRoot.getOfficialMatches(), UserParameter.instance().downloadCurrentMatchlist);
+		// - Integrated matches
+		downloadFilter.checkNode(filterRoot.getSingleMatches(), UserParameter.instance().downloadSingleMatches);
+		downloadFilter.checkNode(filterRoot.getLadderMatches(), UserParameter.instance().downloadLadderMatches);
+		downloadFilter.checkNode(filterRoot.getTournamentGroupMatches(), UserParameter.instance().downloadTournamentGroupMatches);
+		downloadFilter.checkNode(filterRoot.getTournamentPlayoffMatches(), UserParameter.instance().downloadTournamentPlayoffMatches);
+		downloadFilter.checkNode(filterRoot.getDivisionBattleMatches(), UserParameter.instance().downloadDivisionBattleMatches);
+
+		// Team Data
+		downloadFilter.checkNode(filterRoot.getTeamData(), UserParameter.instance().xmlDownload);
+
+		// Series Data (fixtures)
+		downloadFilter.checkNode(filterRoot.getSeriesData(), UserParameter.instance().fixtures);
+
+		normalDownloadPanel.setLayout(new BorderLayout());
+		normalDownloadPanel.add(new JScrollPane(downloadFilter), BorderLayout.CENTER);
+		normalDownloadPanel.setSize(240, 280);
 		normalDownloadPanel.setLocation(10, 10);
 		getContentPane().add(normalDownloadPanel);
 
@@ -166,19 +187,30 @@ public class DownloadDialog extends JDialog implements ActionListener {
 		((JSpinner.DateEditor) m_jsSpinner.getEditor()).getFormat().applyPattern("dd.MM.yyyy");
 		matchArchivePanel.add(m_jsSpinner, BorderLayout.EAST);
 
-		oldFixturePanel.add(matchArchivePanel, BorderLayout.SOUTH);
+		// Show HRF FileDialog
+		m_jchShowSaveDialog.setToolTipText(hov.getLanguageString("tt_Optionen_Show_SaveHRF_Dialog"));
+		m_jchShowSaveDialog.setOpaque(false);
+		m_jchShowSaveDialog.addActionListener(this);
+
+		final JPanel diverseOptionsPanel = new JPanel(new BorderLayout(1,2));
+		diverseOptionsPanel.add(matchArchivePanel, BorderLayout.NORTH);
+		diverseOptionsPanel.add(m_jchShowSaveDialog, BorderLayout.SOUTH);
+
+		oldFixturePanel.add(diverseOptionsPanel, BorderLayout.SOUTH);
 
 		specialDownload.add(oldFixturePanel);
 
-		specialDownload.setSize(300, 200);
-		specialDownload.setLocation(220, 10);
+
+
+		specialDownload.setSize(260, 280);
+		specialDownload.setLocation(260, 10);
 		getContentPane().add(specialDownload);
 
 		m_jbDownload.setToolTipText(hov.getLanguageString("tt_Download_Start"));
 		m_jbDownload.addActionListener(this);
 		m_jbDownload.setFont(m_jbDownload.getFont().deriveFont(Font.BOLD));
 		m_jbDownload.setSize(140, 30);
-		m_jbDownload.setLocation(10, 220);
+		m_jbDownload.setLocation(10, 300);
 		InputMap buttonKeys = m_jbDownload.getInputMap(JButton.WHEN_FOCUSED);
 		buttonKeys.put(KeyStroke.getKeyStroke(KeyEvent.VK_ENTER,0,false), "pressed");
 		buttonKeys.put(KeyStroke.getKeyStroke(KeyEvent.VK_ENTER,0,true), "released");
@@ -189,17 +221,17 @@ public class DownloadDialog extends JDialog implements ActionListener {
 		m_jbProxy.addActionListener(this);
 		m_jbProxy.setFont(m_jbProxy.getFont().deriveFont(Font.BOLD));
 		m_jbProxy.setSize(140, 30);
-		m_jbProxy.setLocation(195, 220);
+		m_jbProxy.setLocation(195, 300);
 
 		getContentPane().add(m_jbProxy);
 
 		m_jbAbort.setToolTipText(hov.getLanguageString("tt_Download_Abbrechen"));
 		m_jbAbort.addActionListener(this);
 		m_jbAbort.setSize(140, 30);
-		m_jbAbort.setLocation(380, 220);
+		m_jbAbort.setLocation(380, 300);
 		getContentPane().add(m_jbAbort);
 
-		setSize(530, 280);
+		setSize(530, 360);
 
 		final Dimension size = getToolkit().getScreenSize();
 
@@ -227,10 +259,24 @@ public class DownloadDialog extends JDialog implements ActionListener {
 		boolean bOK = true;
 		HOModel model = hov.getModel();
 		int teamId = model.getBasics().getTeamId();
-		
+
+		// Save Dialog's settings
+		// the chpp-api only allows to download all current matches, the dialog settings only filters the match
+		// types which of them should be stored in HO database. The settings are interpreted elsewhere.
+		// Dialogs node "Current Matches" is checked, if any of the following is checked
+		UserParameter.instance().downloadCurrentMatchlist = downloadFilter.isChecked(filterRoot.getOfficialMatches());
+		UserParameter.instance().downloadDivisionBattleMatches = downloadFilter.isChecked(filterRoot.getDivisionBattleMatches());
+		UserParameter.instance().downloadTournamentPlayoffMatches = downloadFilter.isChecked(filterRoot.getTournamentPlayoffMatches());
+		UserParameter.instance().downloadTournamentGroupMatches = downloadFilter.isChecked(filterRoot.getTournamentGroupMatches());
+		UserParameter.instance().downloadLadderMatches = downloadFilter.isChecked(filterRoot.getLadderMatches());
+		UserParameter.instance().downloadSingleMatches = downloadFilter.isChecked(filterRoot.getSingleMatches());
+
+		UserParameter.instance().showHRFSaveDialog = m_jchShowSaveDialog.isSelected();
+		UserParameter.instance().xmlDownload = m_jchHRF.isSelected();
+		UserParameter.instance().fixtures = m_jchFixtures.isSelected();
+
 		// Always test that teamId exists, it won't on the very first download
-		
-		if (m_jchOwnFixtures.isSelected() && (teamId > 0)) {
+		if ( this.downloadFilter.isChecked(filterRoot.getCurrentMatches())  && (teamId > 0)) {
 			// Only get lineups for own fixtures
 			bOK = (OnlineWorker.getMatches(teamId, false, true, true) != null);
 			if (bOK) {
@@ -251,28 +297,23 @@ public class DownloadDialog extends JDialog implements ActionListener {
 		}
 
 		if (bOK && m_jchOldFixtures.isSelected() && (teamId > 0)) {
-			if (m_jlOldSeasons.getSelectedValues() != null) {
-				final Object[] saisons = m_jlOldSeasons.getSelectedValues();
-				for (int i = 0; i < saisons.length; i++) {
-					if (saisons[i] instanceof CBItem) {
-						// Liga
-						final int saisonid = ((CBItem)saisons[i]).getId();
+			for (Object s : m_jlOldSeasons.getSelectedValuesList()) {
+				if (s instanceof CBItem) {
+					final int seasonId = ((CBItem) s).getId();
+					// Abfragen!
+					final LigaAuswahlDialog auswahlDialog = new LigaAuswahlDialog(this, seasonId);
+					final int leagueId = auswahlDialog.getLigaID();
 
-						// Abfragen!
-						final LigaAuswahlDialog auswahlDialog = new LigaAuswahlDialog(this, saisonid);
-						final int ligaid = auswahlDialog.getLigaID();
-
-						if (ligaid > -2) {
-							bOK = OnlineWorker.getSpielplan(saisonid, ligaid);
-						}
-						if (!bOK) {
-							break;
-						}
+					if (leagueId > -2) {
+						bOK = OnlineWorker.getSpielplan(seasonId, leagueId);
+					}
+					if (!bOK) {
+						break;
 					}
 				}
 			}
 		}
-		
+
 		// Lastly, so that the matches for training are there
 		if (bOK && m_jchHRF.isSelected()) {
 			OnlineWorker.getHrf(this);