+++ src/main/java/core/option/DownloadPanel.java
@@ -15,7 +15,9 @@ import javax.swing.event.ChangeEvent;
 
 
 /**
- * Checkboxes in Download Dialog checked or not 
+ * Checkboxes in Download Dialog checked or not
+ *
+ * (DownloadPanel is not used anymore - the download options are now stored in DownloadDialog)
  */
 public final class DownloadPanel extends ImagePanel
     implements javax.swing.event.ChangeListener, java.awt.event.ItemListener
@@ -42,7 +44,7 @@ public final class DownloadPanel extends ImagePanel
     public final void itemStateChanged(java.awt.event.ItemEvent itemEvent) {        
         core.model.UserParameter.temp().xmlDownload = m_jchXMLDownload.isSelected();
         core.model.UserParameter.temp().fixtures = m_jchFixtures.isSelected();
-        core.model.UserParameter.temp().currentMatchlist = m_jchCurrentMatchlist.isSelected();
+        core.model.UserParameter.temp().downloadCurrentMatchlist = m_jchCurrentMatchlist.isSelected();
         core.model.UserParameter.temp().showHRFSaveDialog = m_jchShowSaveDialog.isSelected();
     }
 
@@ -70,7 +72,7 @@ public final class DownloadPanel extends ImagePanel
 		m_jchCurrentMatchlist = new JCheckBox(core.model.HOVerwaltung.instance().getLanguageString("download.currentmatches"));
 		m_jchCurrentMatchlist.setToolTipText(core.model.HOVerwaltung.instance().getLanguageString("download.currentmatches.tt"));
 		m_jchCurrentMatchlist.setOpaque(false);
-		m_jchCurrentMatchlist.setSelected(core.model.UserParameter.temp().currentMatchlist);
+		m_jchCurrentMatchlist.setSelected(core.model.UserParameter.temp().downloadCurrentMatchlist);
 		m_jchCurrentMatchlist.addItemListener(this);
 		placement.gridy = 1;
 		add(m_jchCurrentMatchlist, placement);